requirements.txt
opencv-python
numpy
paho-mqtt